src/requirements.txt
fastapi[all]==0.96.0
SQLAlchemy==2.0.15
alembic==1.11.1
asyncpg==0.27.0
psycopg2-binary==2.9.6
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
pytest==7.3.2
pytest-asyncio==0.21.0
httpx==0.24.1
